requirements-stdborg.txt
aiofiles
opencv-contrib-python
aiohttp
aria2p
beautifulsoup4
coffeehouse>=2.1.0
cowpy
cfscrape
cryptg
emoji
googletrans
google_images_download>=2.7.1
google-api-python-client
google_auth_oauthlib
gsearch>=1.6.0
gtts
hachoir
Pillow>=5.3.0
psycopg2
pySmartDL
python-magic
qrcode
regex
requests>=2.18.4
PyGithub
speedtest-cli
sqlalchemy>=1.3.13
telegraph
urbandict
wikipedia
youtube-dl
lxml
heroku3 
git+https://github.com/LonamiWebs/Telethon
git+https://github.com/ahupp/python-magic
oauth2client
httplib2>=0.15.0
humanize
dnspython
numpy>=1.11.1
databases[sqlite]
praw
opencv_python>=1.1.1
spamwatch>=0.2.0
aioschedule
REPO_LINK
feedparser
apscheduler
setuptools>=40.3.0
lyricsgenius
jotquote
csvfaker
pytube3
gitpython
search-engine-parser>=0.4.2
wget
hurry.filesize
pymongo
telethon-session-sqlalchemy>=0.2.6
TgCrypto>=1.2.0
